new draw calls, size_t in planes, fixing indexing
 - shiny::signal_draw can now be used to setup full-pipelined
   draws, so that everything needed for the draw is specified in
   one spot, DX12 styles.

 - made a bunch of uint become size_t

 - index16/index32 introduced, to much yawning.

+++ include/shiny/data_declaration.hpp
@@ -13,14 +13,14 @@ namespace shiny
 		data_declaration_t(data_declaration_t&&) = delete;
 
 		auto streams() const -> data_streams_t const&;
-		auto stride() const -> uint;
+		auto stride() const -> size_t;
 
 	private:
 		data_declaration_t(data_streams_t const&);
 
 	private:
 		data_streams_t streams_;
-		uint stride_;
+		size_t stride_;
 
 		friend struct runtime_t;
 	};
