+++ include/shiny/platform/win32/generic_buffer.hpp
@@ -8,7 +8,7 @@ namespace shiny
 	struct generic_buffer_t : shiny::buffer_t
 	{
 		generic_buffer_t(context_ptr const& ctx, buffer_usage_t usage, element_format_t format, uint elements, void const* data, uint data_elemcount);
-		generic_buffer_t(context_ptr const& ctx, buffer_usage_t usage, uint stride, uint elements, void const* data, uint data_elemcount);
+		generic_buffer_t(context_ptr const& ctx, buffer_usage_t usage, size_t stride, uint elements, void const* data, uint data_elemcount);
 
 		// todo: constructor that shares the data of another genric-buffer but has a different view
 
@@ -30,7 +30,7 @@ namespace shiny
 	}
 
 	inline auto create_generic_buffer(
-		context_ptr const& ctx, buffer_usage_t usage, uint stride,
+		context_ptr const& ctx, buffer_usage_t usage, size_t stride,
 		uint elements, void const* data, uint data_elemcount)
 		-> generic_buffer_ptr
 	{
